Move botan integration in to separate namespace
